requirements.txt
flask>=3.0.0
numpy==2.2.6
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.8
scipy>=1.10
seaborn>=0.13
tensorflow==2.20.0
keras==3.12.1
pillow>=10.0
opencv-python>=4.8
